# Conflicts:
#	src/Controleur/MoteurJeu.java
#	src/Global/Config.java
#	src/IHM/Console/IHMConsole.java
#	src/IHM/Graphique/IHMGraphique.java
#	src/Modele/Coups/CoupDeplacement.java
#	src/Modele/Jeu/JeuGraphe.java
#	src/Modele/Jeux/Jeu.java
#	src/Modele/Jeux/JeuConcret.java

+++ src/Modele/IA/H1.java
@@ -1,12 +1,12 @@
 package Modele.IA;
 
-import Modele.Jeu.Jeu;
+import Modele.Jeux.Jeu;
 
 import java.util.ArrayList;
 
 public class H1 implements Heuristique {
     @Override
-    public int evaluer(Jeu j, int pdvJoueur) {
+    public double evaluer(Jeu j, int pdvJoueur) {
         if (j.estTermine()) {
             ArrayList<Integer> winners = j.getWinner();
             if (winners.contains(pdvJoueur))
